+++ src/Products/PloneMeeting/monkey.py
@@ -3,14 +3,9 @@ from Acquisition import aq_base
 from cPickle import dumps
 from imio.helpers.cache import get_current_user_id
 from imio.helpers.cache import get_plone_groups_for_user
-from imio.helpers.security import fplog
 from plone import api
 from plone.app.querystring import queryparser
 from plone.memoize import ram
-from plone.restapi.deserializer import boolean_value
-from plone.restapi.deserializer import json_body
-from plone.restapi.services import Service
-from plonemeeting.restapi import logger as pmrestapi_logger
 from Products.Archetypes.BaseObject import BaseObject
 from Products.Archetypes.Field import Field
 from Products.PloneMeeting import logger
@@ -24,9 +19,6 @@ from types import StringType
 from z3c.form import interfaces
 from z3c.form.widget import SequenceWidget
 from zope.ramcache.ram import Storage
-
-import json
-import os
 
 
 def _patched_equal(context, row):
@@ -183,44 +175,6 @@ def getMemberInfo(self, memberId=None):
 
 MembershipTool.getMemberInfo = getMemberInfo
 logger.info("Monkey patching Products.PlonePAS.tools.membership.MembershipTool (getMemberInfo)")
-
-
-# plonemeeting.restapi, need to monkeypatch here because order of packages
-# monkey patching in plonemeeting.restapi does not seem to work...
-
-Service.__old_pm_render = Service.render
-
-
-def render(self):
-    """Monkeypatched to add fplog."""
-    query_string = self.request.get('QUERY_STRING', '')
-    extras = 'name={0} url={1}{2}'.format(
-        self.__name__,
-        self.request.get('ACTUAL_URL'),
-        query_string and " query_string={0}".format(query_string) or '')
-    fplog("restapi_call", extras=extras)
-
-    # debug may be enabled by passing debug=true as parameter to the restapi call
-    # or when setting the RESTAPI_DEBUG environment variable
-    debug = boolean_value(self.request.form.get('debug', False)) or \
-        boolean_value(os.environ.get('RESTAPI_DEBUG', False)) or \
-        "debug=true" in query_string  # with POST, URL parameters are not in self.request.form
-    # log the input when debug is enabled
-    if debug:
-        # with POST, data is in the body
-        if self.request.get('method', 'GET') != 'GET':
-            data = json_body(self.request)
-            fplog("restapi_call_debug",
-                  extras="INPUT: \n" + json.dumps(data, indent=4, sort_keys=True))
-    res = self.__old_pm_render()
-    # log the output when debug is enabled
-    if debug:
-        fplog("restapi_call_debug", extras="OUTPUT: \n" + res)
-    return res
-
-
-Service.render = render
-pmrestapi_logger.info("Monkey patching plone.restapi.services.RestService (render)")
 
 
 Storage.__old_pm_getEntry = Storage.getEntry